requirements.txt
alembic==0.9.3
blinker==1.4
click==7.1.2
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-PageDown==0.4.0
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.7.3
python-editor==1.0.3
six==1.12
SQLAlchemy==1.1.11
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
Flask-SSLify==0.1.5
psycopg2-binary==2.9.4
python-dotenv==0.21.0
bleach==5.0.1
Markdown==3.4.1
Flask-HTTPAuth==4.7.0
fake-useragent==0.1.11
coverage==6.5.0
gunicorn==19.7.1
pymysql==0.7.11
redis==3.5.3
celery==5.1.2
flower==1.2.0
